refactor: Fix design system violations and improve code quality
Addressed all critical formatting and design system issues identified by
formatting-cop and design-system-cop agents.

**Critical Fixes:**
- Fixed inline RGB colors in PointPoolDisplay (bypassed design tokens)
- Removed console.error debug statements from production code
- Fixed ESLint warnings (unused variables, img tags)

**Design Token Conversions:**
- Converted 100+ primitive red classes to semantic `destructive` tokens
  - Error messages: text-red-500 → text-destructive
  - Error backgrounds: bg-red-200 → bg-destructive/10
  - Error borders: border-red-500 → border-destructive
- Converted 100+ primitive blue classes to semantic tokens
  - Primary buttons: bg-blue-500 → bg-primary
  - Info boxes: bg-blue-50 → bg-muted
  - Focus rings: focus:ring-blue-500 → focus:ring-primary
  - Selected states: bg-blue-100 → bg-primary/10

**File Size Refactoring:**
- Reduced CharacterCreationWizard.tsx from 491 → 287 lines (42% reduction)
- Extracted wizard state management to useCharacterCreationWizard hook
- Extracted point pool logic to useCharacterPointPools hook
- Extracted character finalization to characterFinalization utility

**Files Modified:** 31 components, 3 new files created
**Tests:** All 1,622 tests passing
**Linting:** ESLint and Stylelint clean

This ensures full compliance with the project's 23-color design token system
and file size standards while maintaining all existing functionality.
