+++ src/components/CharacterCreationWizard/steps/PortraitStep.tsx
@@ -174,7 +174,7 @@ export function PortraitStep({ data, onUpdate, worldConfig }: PortraitStepProps)
             type="button"
             onClick={handleGeneratePortrait}
             disabled={isGenerating}
-            className="px-4 py-2 bg-blue-500 text-white rounded hover:bg-blue-700 disabled:opacity-50 disabled:cursor-not-allowed"
+            className="px-4 py-2 bg-primary text-white rounded hover:bg-primary/90 disabled:opacity-50 disabled:cursor-not-allowed"
           >
             Generate Portrait
           </button>
