+++ src/components/WorldList/WorldList.tsx
@@ -28,23 +28,23 @@ const WorldList: React.FC<WorldListProps> = ({ worlds, currentWorldId, onSelectW
   if (worlds.length === 0) {
     return (
       <section data-testid="world-list-empty-message" className="p-12 text-center bg-gray-100 rounded-lg">
-        <div className="w-24 h-24 bg-blue-100 rounded-full mx-auto mb-4 flex items-center justify-center">
-          <Globe className="w-12 h-12 text-blue-700" aria-hidden="true" />
+        <div className="w-24 h-24 bg-primary/10 rounded-full mx-auto mb-4 flex items-center justify-center">
+          <Globe className="w-12 h-12 text-primary" aria-hidden="true" />
         </div>
         <h2 className="text-2xl font-semibold text-gray-700 mb-2">Welcome to Narraitor!</h2>
         <p className="text-gray-700 mb-2">Begin your storytelling journey by creating your first world.</p>
         <p className="text-gray-500 text-sm mb-6">Each world is a unique setting with its own rules, attributes, and possibilities.</p>
-        <div className="bg-blue-50 border border-blue-200 rounded-lg p-4 max-w-md mx-auto">
-          <p className="text-sm text-blue-900 font-semibold mb-3">
+        <div className="bg-muted border border-border rounded-lg p-4 max-w-md mx-auto">
+          <p className="text-sm text-muted-foreground font-semibold mb-3">
             Getting Started Guide:
           </p>
-          <ol className="text-sm text-blue-700 text-left list-decimal list-inside space-y-2">
+          <ol className="text-sm text-muted-foreground text-left list-decimal list-inside space-y-2">
             <li><strong>Create a World</strong> - Define your setting, theme, and game rules</li>
             <li><strong>Build Characters</strong> - Populate your world with unique personalities</li>
             <li><strong>Start Playing</strong> - Begin your interactive narrative experience</li>
           </ol>
-          <div className="mt-4 pt-3 border-t border-blue-200">
-            <p className="text-xs text-blue-700">
+          <div className="mt-4 pt-3 border-t border-border">
+            <p className="text-xs text-muted-foreground">
               <strong>Tip:</strong> You can create multiple worlds and switch between them anytime!
             </p>
           </div>
