+++ src/components/shared/WorldFormFields/WorldFormFields.tsx
@@ -31,19 +31,19 @@ export function WorldNameInput({
   return (
     <div className={className}>
       <Label htmlFor="world-name" className="block text-sm font-medium text-gray-700 mb-2">
-        World Name {required && <span className="text-red-500">*</span>}
+        World Name {required && <span className="text-destructive">*</span>}
       </Label>
       <Input
         id="world-name"
         type="text"
-        className="w-full px-3 py-2 border border-gray-300 rounded-md focus:outline-none focus:ring-2 focus:ring-blue-500"
+        className="w-full px-3 py-2 border border-gray-300 rounded-md focus:outline-none focus:ring-2 focus:ring-primary"
         placeholder={placeholder || "E.g., Neo-Tokyo..."}
         value={value}
         onChange={(e) => onChange(e.target.value)}
         disabled={disabled}
       />
       {error && (
-        <p className="text-sm text-red-500 mt-1">{error}</p>
+        <p className="text-sm text-destructive mt-1">{error}</p>
       )}
     </div>
   );
@@ -71,11 +71,11 @@ export function GenreSelect({
   return (
     <div className={className}>
       <Label htmlFor="world-genre" className="block text-sm font-medium text-gray-700 mb-2">
-        Genre {required && <span className="text-red-500">*</span>}
+        Genre {required && <span className="text-destructive">*</span>}
       </Label>
       <Select
         id="world-genre"
-        className="w-full px-3 py-2 border border-gray-300 rounded-md focus:outline-none focus:ring-2 focus:ring-blue-500"
+        className="w-full px-3 py-2 border border-gray-300 rounded-md focus:outline-none focus:ring-2 focus:ring-primary"
         value={value}
         onChange={(e) => onChange(e.target.value)}
         disabled={disabled}
@@ -88,7 +88,7 @@ export function GenreSelect({
         ))}
       </Select>
       {error && (
-        <p className="text-sm text-red-500 mt-1">{error}</p>
+        <p className="text-sm text-destructive mt-1">{error}</p>
       )}
     </div>
   );
@@ -125,25 +125,25 @@ export function DescriptionTextArea({
   return (
     <div className={className}>
       <Label htmlFor="world-description" className="block text-sm font-medium text-gray-700 mb-2">
-        {label} {required && <span className="text-red-500">*</span>}
+        {label} {required && <span className="text-destructive">*</span>}
       </Label>
       <Textarea
         id="world-description"
         rows={rows}
-        className="w-full px-3 py-2 border border-gray-300 rounded-md focus:outline-none focus:ring-2 focus:ring-blue-500"
+        className="w-full px-3 py-2 border border-gray-300 rounded-md focus:outline-none focus:ring-2 focus:ring-primary"
         placeholder={placeholder}
         value={value}
         onChange={(e) => onChange(e.target.value)}
         disabled={disabled}
         maxLength={maxLength}
       />
-      
+
       {/* Character count */}
       {maxLength && (
         <div className="flex justify-between items-center mt-1">
           <div className="flex-1">
             {error && (
-              <p className="text-sm text-red-500">{error}</p>
+              <p className="text-sm text-destructive">{error}</p>
             )}
           </div>
           <p className={`text-xs ${isNearLimit ? 'text-amber-500' : 'text-gray-500'}`}>
@@ -151,10 +151,10 @@ export function DescriptionTextArea({
           </p>
         </div>
       )}
-      
+
       {/* Error without character count */}
       {!maxLength && error && (
-        <p className="text-sm text-red-500 mt-1">{error}</p>
+        <p className="text-sm text-destructive mt-1">{error}</p>
       )}
     </div>
   );
@@ -203,24 +203,24 @@ export function ErrorDisplay({
   if (!error) return null;
 
   return (
-    <div className={`bg-red-200 border border-red-500 rounded-md p-4 ${className}`}>
+    <div className={`bg-destructive/10 border border-destructive rounded-md p-4 ${className}`}>
       <div className="flex">
         <div className="flex-1">
-          <h3 className="text-sm font-medium text-red-700">
+          <h3 className="text-sm font-medium text-destructive">
             Error
           </h3>
-          <p className="text-sm text-red-700 mt-1">
+          <p className="text-sm text-destructive mt-1">
             {error}
           </p>
         </div>
-        
+
         {(onRetry || onDismiss) && (
           <div className="flex space-x-2 ml-4">
             {onRetry && (
               <Button
                 onClick={onRetry}
                 variant="ghost"
-                className="text-sm font-medium text-red-700 hover:text-red-900 underline"
+                className="text-sm font-medium text-destructive hover:text-destructive/80 underline"
               >
                 Retry
               </Button>
@@ -229,7 +229,7 @@ export function ErrorDisplay({
               <Button
                 onClick={onDismiss}
                 variant="ghost"
-                className="text-sm font-medium text-red-700 hover:text-red-900"
+                className="text-sm font-medium text-destructive hover:text-destructive/80"
               >
                 ×
               </Button>
