+++ src/components/ui/toast/toast.tsx
@@ -42,7 +42,7 @@ export interface ToastProps extends React.HTMLAttributes<HTMLDivElement> {
 
 const toastVariants = {
   success: "bg-green-200 border-green-500 text-green-700",
-  error: "bg-red-200 border-red-500 text-red-700",
+  error: "bg-destructive/10 border-destructive text-destructive",
   warning: "bg-amber-200 border-amber-500 text-amber-700",
   info: "bg-blue-100 border-blue-300 text-blue-700",
 }
