+++ src/components/world/SkillEditor/SkillEditor.tsx
@@ -214,7 +214,7 @@ export function SkillEditor({
       {errors.length > 0 && (
         <div className="space-y-2">
           {errors.map((error, index) => (
-            <div key={index} className="text-sm text-red-500" role="alert">
+            <div key={index} className="text-sm text-destructive" role="alert">
               {error}
             </div>
           ))}
@@ -224,7 +224,7 @@ export function SkillEditor({
       <form onSubmit={handleSubmit} role="form" className="space-y-6">
         <div className="space-y-4">
           <div>
-            <Label htmlFor="skill-name">Skill Name <span className="text-red-500">*</span></Label>
+            <Label htmlFor="skill-name">Skill Name <span className="text-destructive">*</span></Label>
             <Input
               id="skill-name"
               type="text"
@@ -237,7 +237,7 @@ export function SkillEditor({
           </div>
 
           <div>
-            <Label htmlFor="skill-description">Description <span className="text-red-500">*</span></Label>
+            <Label htmlFor="skill-description">Description <span className="text-destructive">*</span></Label>
             <Textarea
               id="skill-description"
               value={formData.description || ''}
